Moved the new_rule_based tagger into rule_based
