+++ tests/data/taggers/rule_based/mwe_lexicon.tsv
