+++ tests/data/taggers/rule_based/rule_based_mwe_input_output.json
