+++ tests/data/taggers/rule_based/rule_based_single_input_output.json
