+++ tests/data/taggers/rule_based/single_lexicon.tsv
